==> notebooks/01-data-preprocessing.ipynb <==
import pandas as pd

# Read the dataframe 
df = pd.read_csv("/Users/admin/PycharmProjects/Cocktail-Recommendation-System/database/final_cocktails.csv")
df.head(5)

df.describe()

df.info()

df.isnull().sum()

# drop the Unnamed column
df.drop(columns=["Unnamed: 0"], inplace=True)
df.head(5)

# Check the "ingredients" column
print(type(df["ingredients"][2]))
print(type(df["ingredientMeasures"][3]))


print(df["ingredients"].head(3))

import ast
df['ingredients'] = df['ingredients'].apply(ast.literal_eval)
df['ingredientMeasures'] = df['ingredientMeasures'].apply(ast.literal_eval)

# Check the "ingredients" and "ingredientsMeasure" columns
print(type(df["ingredients"][2]))
print(type(df["ingredientMeasures"][3]))


# Normalize the ingredients column
df['ingredients'] = df['ingredients'].apply(
    lambda ingredients: [ingredient.strip().lower() for ingredient in ingredients]
)

# Display the normalized ingredients
df['ingredients'].head()

# Standardize alcoholic, category, glassType
df["alcoholic"] = df["alcoholic"].str.lower().str.strip()
df["category"] = df["category"].str.lower().str.strip()
df["glassType"] = df["glassType"].str.lower().str.strip()

print(df["alcoholic"].unique())
print(df["category"].unique())
print(df["glassType"].unique())

# The most common ingredients
most_common_ingredients = df['ingredients'].explode().value_counts().head(10) # Transform the list of ingredients into a long format
most_common_ingredients

most_frequent_glass_types = df['glassType'].value_counts().head(10)
most_frequent_glass_types

# number of alcoholic and non-alcoholic cocktails
number_of_alcoholic = df[df['alcoholic'] == 'alcoholic'].shape[0]
number_of_non_alcoholic = df[df['alcoholic'] == 'non-alcoholic'].shape[0]
print(f"Number of alcoholic cocktails: {number_of_alcoholic}")
print(f"Number of non-alcoholic cocktails: {number_of_non_alcoholic}")

# Review the dataframe after cleaning
df.head(5)

# Replace None in the ingredientMeasures column with a placeholder
df['ingredientMeasures'] = df['ingredientMeasures'].apply(
    lambda measures: [measure if measure is not None else 'to taste' for measure in measures]
)
# Display the updated ingredientMeasures column
df['ingredientMeasures'].head(20)

# Check the mis matched lengths of ingredients and ingredientMeasures
mismatched_lengths = df[df.apply(lambda x: len(x['ingredients']) != len(x['ingredientMeasures']), axis=1)]
print(f"Number of mismatched lengths: {mismatched_lengths.shape[0]}")

df.head(10)

# Save the cleaned dataframe to a new CSV file
df.to_csv("/Users/admin/PycharmProjects/Cocktail-Recommendation-System/database/cleaned_cocktails.csv", index=False)
