==> notebooks/02-vectorization.ipynb <==
import ast

# Read the clean CSV
import pandas as pd

df = pd.read_csv('/Users/admin/PycharmProjects/Cocktail-Recommendation-System/database/cleaned_cocktails.csv')
df.head(5)

It is just a demo to under the concept of vectorized database :)))

#Build the ingredient vocabulary
import ast

vocabulary = set()
for ingredients in df['ingredients']:
    ingredients_list = ast.literal_eval(ingredients) # Ex: "[1, 2, 3, 4]" to [1, 2, 3, 4]
    for ingredient in ingredients_list:
        vocabulary.add(ingredient)
vocabulary = sorted(vocabulary)
vocabulary

# Encoding ingredients as vectors
def encode_ingredients(vocabulary, ingredients):
    vector = [0] * len(vocabulary)
    for ingredient in ast.literal_eval(ingredients):
        cleaned_ingredient = ingredient.strip().lower()
        if cleaned_ingredient in vocabulary:
            vector[vocabulary.index(cleaned_ingredient)] = 1
    return vector

df["ingredient_vec"] = df["ingredients"].apply(lambda x: encode_ingredients(vocabulary, x))
df["ingredient_vec"].head(5)

df.head(5)

# Use Binary Encoding for alcoholic column
df["is_alcoholic"] = df["alcoholic"].map({'alcoholic': 1, 'non alcoholic': 0})
df['alcoholic'] = df['is_alcoholic']

df = df.drop(columns='is_alcoholic')

df.head(5)

df.to_csv("/Users/admin/PycharmProjects/Cocktail-Recommendation-System/database/vectorized_cocktails.csv", index=False)

df.info()
